+++ test/fixture/example/invalid-values.sh
@@ -3,6 +3,7 @@ export DEBUG=yes
 export EARTH_ATOM_COUNT=5.9722e24
 export GRPC_TIMEOUT=10S
 export LOG_LEVEL=silly
+export PORT=65536
 export READ_DSN=
 export REDIS_PRIMARY_SERVICE_HOST=.redis.example.org
 export REDIS_PRIMARY_SERVICE_PORT=65536