+++ test/fixture/example/valid-values.sh
@@ -3,6 +3,7 @@ export DEBUG=false
 export EARTH_ATOM_COUNT=5972200000000000000000000
 export GRPC_TIMEOUT=PT0.01S
 export LOG_LEVEL=error
+export PORT=8080
 export READ_DSN="host=localhost dbname=readmodels user=projector"
 export REDIS_PRIMARY_SERVICE_HOST=redis.example.org
 export REDIS_PRIMARY_SERVICE_PORT=6379